% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cFDR.R
\name{cfdr}
\alias{cfdr}
\title{cfdr}
\usage{
cfdr(P_i, P_j, sub = 1:length(P_i))
}
\arguments{
\item{P_i}{vector of p-values for the principal phenotype.}

\item{P_j}{vector of p-values or adjusted p-values for the conditional phenotype. If controls are shared between GWAS, the p-values must be adjusted using the function \code{\link{exp_quantile}}.}

\item{sub}{set of indices; only compute cFDR at this subset of p-value pairs. Typically, only SNPs with low values of P_i or P_j are of interest; it is possible to increase the speed of the procedure by only calculating cFDR at such SNPs.}
}
\description{
Compute estimated cFDR values from a set of pairs of p values, or adjusted p-values if controls are shared.
}
\details{
For a set of p-values or adjusted p values \eqn{P_{j}}{P_i} for the principal phenotype, and a set of p values \eqn{P_{j}}{P_j} for the conditional phenotype, the cFDR for SNP \eqn{x} is estimated as

\deqn{
  \widehat{cFDR}(x)=P_{i}(x) \frac{\textrm{(# SNPs with $P_{j} \leq P_{j}(x)$}{\textrm{(# SNPs with $P_{i} \leq P_{i}(x), P_{j} \leq P_{j}(x))}}}
}{
  est. cFDR(x) = P_i(x) * (# P_j \le P_j(x)) / (# P_i \le P_i(x) AND P_j \le P_j(x))
}

The procedure can be sped up by specifying a subset of SNPs at which to estimate the cFDR. This is most effective if a boundary of the subset can be written as \eqn{P_{j} = f(P_{i})}{P_j = f(P_i)}, where \eqn{f} is non-increasing. Equivalently, if \eqn{S} is the set of all p-value pairs, the subset \eqn{S'} should satisfy the condition

\deqn{
  \forall \left( p_{i}',p_{j}' \in S', p_{i},p_{j} \in S \right) \: p_{i} \leq  p_{i}', p_{j} \leq p_{j}' \implies p_{i},p_{j} \in S'
}{
  for all (p_i',p_j') in S', (p_i,p_j) in S:   p_i \le p_i', q_j \le p_j   =>    (p_i,p_j) in S'.
}

An example of such a subset is the set of \eqn{p_{i},p_{j}}{p_i,p_j} with \eqn{p_{i}^2 + p_{j}^2 < \alpha}{p_i^2 + p_j^2 < \alpha}.

The estimated cFDR is a conservative estimate (upper bound) on the quantity

\deqn{
  Pr(H_{0}^{i} | P_{i}<P_{i}(x),P_{j}<P_{j}(x))
 }{
  Pr(H0 | P_i<P_i(x),P_j<P_j(x))
 }

where \eqn{H_{0}^{i}}{H0} is the null hypothesis at SNP \eqn{x} for the principal phenotype (ie, SNP \eqn{x} is not associated with the principal phenotype)
}
\examples{
require(mnormt)
P = 2*pnorm(-abs(rmnorm(10000,varcov=diag(c(0.8,0.8))+0.2)))
P_i = P[,1]; P_j=P[,2]
W = which((qnorm(P_i/2)^2) + (qnorm(P_j/2)^2) > 4)

# Using function
C = cfdr(P_i,P_j,sub=W); C[W[1]]

# Explicit computation
P_i[W[1]]/ (length(which (P_i <= P_i[1] & P_j <= P_j[1])) / length(which (P_j <= P_j[1])) )
}
\author{
James Liley
}

