% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cfdrCommonControl.R
\docType{package}
\name{cfdrCommonControl}
\alias{cfdrCommonControl}
\alias{cfdrCommonControl-package}
\title{cfdrCommonControl: a package implementing the Bayesian conditional false discovery rate method allowing for common control samples.}
\description{
As the body of GWAS data grows interest has developed in the degree of shared genetic basis (pleiotropy) between clinically similar phenotypes. Shared genetic basis will frequently tag common pathophysiological pathways, and increased understanding of these can lead to new therapeutic targets.
}
\details{
Meta-analysis of separate diseases by combining case groups has several drawbacks, most notably that there is no direct way to assign probability of association to each disease individually. A useful quantity for testing association with a single disease while taking into account the results of a GWAS on a potentially related disease is the Bayesian conditional false discovery rate (cFDR).

This is defined for a SNP with P values \eqn{p_{i}}{p_i}, \eqn{p_{j}}{p_j} in two GWAS \eqn{i} and \eqn{j} as

\deqn{
 cFDR(p_{i},p_{j})=Pr(H_{0}^{(i)}|P_{i} \leq p_{i},P_{j} \leq p_{j})
 }{
 cFDR(p_i,p_j) = Pr(H0i| P_i \le p_i, P_j \le p_j)
 }

where \eqn{H_{0}^{(i)}}{H0i} is the hypothesis that the SNP is not associated with phenotype \eqn{i}, and \eqn{P_{i}}{P_i}, \eqn{P_{j}}{P_j} are random variables representing p-values over all SNPs for phenotypes \eqn{i} and \eqn{j}.

A conservative estimate of the cFDR can be made using summary statistics, implemented in function \code{\link{cfdr}}. If controls are shared between the two GWAS, the p-values for phenotype \eqn{i} must be transformed using functions \code{\link{exp_quantile}}, \code{\link{cor_shared}} to account for expected correlation between summary statistics.

Unlike Benjamini and Hochberg's FDR-controlling procedure, rejecting all SNPs for which estimated cFDR is less than some cutoff \eqn{\alpha} does not limit the overall false discovery rate to \eqn{\alpha}. An upper bound on the false discovery rate can be made using function \code{\link{c2a}}

Further information can be found in

Andreassen OA, Thompson WK, Schork AJ, Ripke S, Mattingsdal M, Kelsoe JR, et al. Improved detection of common variants associated with schizophrenia and bipolar disorder using pleiotropy-informed conditional false discovery rate. PLoS Genetics 2013

Liley AJ, Wallace CW. A pleiotropy-informed Bayesian false discovery rate adapted to a shared control design finds new disease associations from GWAS summary statistics. PLoS Genetics 2015.
}
\section{cfdrCommonControl functions}{

}

