% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cFDR.R
\name{exp_quantile_point}
\alias{exp_quantile_point}
\title{exp_quantile_point}
\usage{
exp_quantile_point(p, pc, rho, pi0 = 1, sigma = 1)
}
\arguments{
\item{p}{Observed p value or set of p values for principal phenotype}

\item{pc}{Observed p value or set of p values for conditional phenotype. Must be of the same length as \code{\link{p}}.}

\item{rho}{Correlation between Z scores induced by shared controls. Output from \code{\link{cor_shared}}.}

\item{pi0}{Parameter for distribution of Z scores for the conditional phenotype (proportion of SNPs null for conditional phenotype).}

\item{sigma}{Parameter for distribution of Z scores for the conditional phenotype (variance of true effect sizes for SNPs non-null for conditional phenotype).}
}
\value{
A list of probabilities of the same length as \code{\link{p}}.
}
\description{
Evaluate the conditional expected quantile of a p value in the shared control design conditional on a p value for a second phenotype
}
\details{
Computes the probability that a p value for a phenotype \eqn{i} is less than some cutoff \code{\link{p}} given that the p value for a second phenotype \eqn{j} is equal to a value \code{\link{pc}}, under the null hypothesis for phenotype \eqn{i}.

If the GWAS for phenotypes \eqn{i} and \eqn{j} share a number of controls, effect sizes for the two phenotypes will be correlated, even at null SNPs. This leads to dependence of the expected quantile both on the number of shared controls and study sizes (through paramter \code{\link{rho}}), and the distribution of Z scores for phenotype \eqn{j} across all SNPs, not all of which will necessarily be null for phenotype \eqn{j}.

The 'true' Z scores (Z scores which we would observe if the observed allele frequencies exactly matched the population allele frequencies) are assumed to follow a mixture distribution of 0 with probability \code{\link{pi0}}, and N(0,\code{\link{sigma}}^2) with probability 1-\code{\link{pi0}}. These parameters can be estimated from an observed distribution of Z scores using the function \code{\link{fit.em}} .

The value of the function is the first derivative of the function \code{\link{exp_quantile}} with respect to \eqn{P_[j]}{P_j}.
}
\examples{
p <- 0.2; pc <- 0.1;
rho <- 0.25
sigma <- 5; pi0 <- 0.95
exp_quantile_point(p,pc,rho,pi0=pi0,sigma=sigma)
# Generally the expected quantile is close to p.
}
\author{
James Liley
}

