% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFDR.R
\name{fit.em}
\alias{fit.em}
\title{fit.em}
\usage{
fit.em(Z, pi0_init = 0.9, sigma_init = 1, tol = 1e-04, verbose = TRUE,
  maxit = 10000)
}
\arguments{
\item{Z}{numeric vector of observed data}

\item{pi0_init}{initial value for \code{\link{pi0}}}

\item{sigma_init}{initial value for \code{\link{sigma}}}

\item{tol}{how small a change lhood prompts continued optimization}

\item{maxit}{maximum number of iterations}
}
\value{
a list containing fitted \code{\link{pi0}}, fitted \code{\link{sigma}}, and a record of fitted parameters at eachs stage of the E-M algorithm.
}
\description{
Fit a specific two Guassian mixture distribution to a set of Z values.
}
\details{
Assumes 
Z ~ N(0,1) with probability  \code{\link{pi0}}, Z ~ N(0,1 + \code{\link{sigma}}^2) with probability 1-\code{\link{pi0}}

We define 'true' Z scores as the Z scores that would be obtained if MAF for both groups exactly matched the corresponding MAFs in the population, or equivalently the expected values of Z scores. If 'true' Z scores are distributed following a 'spike and tail' model of 0 with probability \code{\link{pi0}} and N(0,\code{\link{sigma}}^2) with probability 1-\code{\link{pi0}}, then observed Z scores follow the above distribution.
}
\examples{
sigma <- 2
pi0 <- 0.8
n <- 10000; n0=round(pi0*n); n1=n-n0
Z <- c(rnorm(n0,0,1),rnorm(n1,0,sqrt(1+ (sigma^2))))
fit<-fit.em(Z)
fit$pi0
fit$sigma
fit$history
}
\author{
Chris Wallace, James Liley
}

